refactor: move Go2JoystickRoughFull task as  Go2JoystickRough task
