--- conf/ppo/task/go2_joystick_rough_full/mujoco.yaml
@@ -1,109 +1,0 @@
-# @package _global_
-training:
-  task_name: Go2JoystickRoughFull
-  sim_backend: mujoco
-  play_steps: 2500
-  play_env_num: 16
-  render_spacing: 0.0
-  cam_tracking: true
-  cam_tracking_env_idx: 0
-  cam_tracking_extra_envs: 9
-
-interactive:
-  action_mode: policy
-  policy_obs_mode: auto
-  camera_follow_body: true
-  use_env_visual_model: false
-
-algo:
-  num_envs: 4096
-  num_steps_per_env: 24
-  max_iterations: 2000
-  empirical_normalization: false
-  obs_groups:
-    actor:
-      - actor
-    critic:
-      - critic
-  policy:
-    init_noise_std: 1.0
-  algorithm:
-    learning_rate: 1.0e-3
-    entropy_coef: 1.0e-2
-    
-env:
-  control_config:
-    action_scale: 0.25
-    hip_action_scale: 0.125
-    non_hip_action_scale: 0.25
-    clip_actions: 100.0
-  commands:
-    vel_limit: [[-1.0, -1.0, -1.0], [1.0, 1.0, 1.0]]
-    resampling_time: 10.0
-    heading_command: true
-    heading_range: [-3.141592653589793, 3.141592653589793]
-  terrain_curriculum:
-    enabled: false
-  terrain_generator:
-    seed: 42
-    curriculum: true
-    size: [8.0, 8.0]
-    num_rows: 7
-    num_cols: 7
-    border_width: 10.0
-  terrain_scan:
-    enabled: true
-    geom_name: floor
-  termination_config:
-    terrain_out_of_bounds: true
-    terrain_distance_buffer: 3.0
-  domain_rand:
-    randomize_base_mass: true
-    added_mass_range: [-1.0, 3.0]
-    random_com: true
-    randomize_kp: true
-    kp_multiplier_range: [0.5, 2.0]
-    randomize_kd: true
-    kd_multiplier_range: [0.5, 2.0]
-    push_robots: true
-    push_interval: 625
-    max_force: [1.0, 1.0, 0.5]
-
-reward:
-  scales:
-    lin_vel_z: -2.0
-    ang_vel_xy: -0.05
-    joint_torques_l2: -2.5e-5
-    joint_acc_l2: -2.5e-7
-    joint_pos_limits: -5.0
-    joint_power: -2.0e-5
-    stand_still: -2.0
-    joint_pos_penalty: -1.0
-    joint_mirror: -0.05
-    action_rate: -0.01
-    undesired_contacts: -1.0
-    contact_forces: -1.5e-4
-    tracking_lin_vel: 3.0
-    tracking_ang_vel: 1.5
-    feet_air_time: 0.5
-    feet_air_time_variance: -1.0
-    feet_contact_without_cmd: 0.1
-    feet_slide: -0.1
-    feet_height_body: -5.0
-    feet_gait: 0.5
-    upward: 1.0
-  tracking_sigma: 0.25
-  base_height_target: 0.33
-  stand_still_command_threshold: 0.1
-  joint_pos_penalty_stand_still_scale: 5.0
-  joint_pos_penalty_velocity_threshold: 0.5
-  joint_pos_penalty_command_threshold: 0.1
-  contact_threshold: 1.0
-  contact_forces_threshold: 100.0
-  feet_air_time_threshold: 0.5
-  feet_height_body_target: -0.2
-  feet_height_body_tanh_mult: 2.0
-  feet_gait_std: 0.7071067811865476
-  feet_gait_max_err: 0.2
-  feet_gait_velocity_threshold: 0.5
-  feet_gait_command_threshold: 0.1
