--- src/unilab/assets/robots/go2/scene_rough.xml
@@ -1,58 +1,0 @@
-<mujoco model="go2 rough scene">
-  <include file="go2.xml"/>
-
-  <statistic center="0 0 0.1" extent="0.8"/>
-
-  <visual>
-    <headlight diffuse="0.6 0.6 0.6" ambient="0.3 0.3 0.3" specular="0 0 0"/>
-    <rgba haze="0.15 0.25 0.35 1"/>
-    <global azimuth="-130" elevation="-20"/>
-    <quality offsamples="4"/>
-    <map force="0.01"/>
-  </visual>
-
-  <asset>
-    <texture type="skybox" builtin="gradient" rgb1="0.3 0.5 0.7" rgb2="0 0 0" width="512" height="3072"/>
-    <texture type="2d" name="groundplane" builtin="checker" mark="edge" rgb1="0.2 0.3 0.4" rgb2="0.1 0.2 0.3"
-      markrgb="0.8 0.8 0.8" width="300" height="300"/>
-    <material name="groundplane" texture="groundplane" texuniform="true" texrepeat="5 5" reflectance="0.2"/>
-    <hfield name="terrain_hfield" file="../../hfields/hfield.png" size="29 29 2.74 1.4" />
-  </asset>
-
-  <worldbody>
-    <light pos="0 0 8.0" dir="0 0 -1" directional="true" castshadow="false"/>
-    <geom name="floor" type="hfield" hfield="terrain_hfield" material="groundplane" contype="2" conaffinity="1" priority="1" friction="1.0"/>
-  </worldbody>
-
-  <sensor>
-    <contact name="FL_foot_contact" geom1="floor" geom2="FL" data="found" num="1" reduce="mindist"/>
-    <contact name="FR_foot_contact" geom1="floor" geom2="FR" data="found" num="1" reduce="mindist"/>
-    <contact name="RL_foot_contact" geom1="floor" geom2="RL" data="found" num="1" reduce="mindist"/>
-    <contact name="RR_foot_contact" geom1="floor" geom2="RR" data="found" num="1" reduce="mindist"/>
-    <!-- base -->
-    <contact name="base1_contact" geom1="floor" geom2="base1" data="found" num="1" reduce="mindist"/>
-    <contact name="base2_contact" geom1="floor" geom2="base2" data="found" num="1" reduce="mindist"/>
-    <contact name="base3_contact" geom1="floor" geom2="base3" data="found" num="1" reduce="mindist"/>
-    <!-- hip -->
-    <contact name="FL_hip_contact" geom1="floor" geom2="FL_hip_geom" data="found" num="1" reduce="mindist"/>
-    <contact name="FR_hip_contact" geom1="floor" geom2="FR_hip_geom" data="found" num="1" reduce="mindist"/>
-    <contact name="RL_hip_contact" geom1="floor" geom2="RL_hip_geom" data="found" num="1" reduce="mindist"/>
-    <contact name="RR_hip_contact" geom1="floor" geom2="RR_hip_geom" data="found" num="1" reduce="mindist"/>
-    <!-- thigh -->
-    <contact name="FL_thigh_contact" geom1="floor" geom2="FL_thigh_geom" data="found" num="1" reduce="mindist"/>
-    <contact name="FR_thigh_contact" geom1="floor" geom2="FR_thigh_geom" data="found" num="1" reduce="mindist"/>
-    <contact name="RL_thigh_contact" geom1="floor" geom2="RL_thigh_geom" data="found" num="1" reduce="mindist"/>
-    <contact name="RR_thigh_contact" geom1="floor" geom2="RR_thigh_geom" data="found" num="1" reduce="mindist"/>
-    <!-- calf -->
-    <contact name="FL_calf_contact1" geom1="floor" geom2="FL_calf_geom1" data="found" num="1" reduce="mindist"/>
-    <contact name="FR_calf_contact1" geom1="floor" geom2="FR_calf_geom1" data="found" num="1" reduce="mindist"/>
-    <contact name="RL_calf_contact1" geom1="floor" geom2="RL_calf_geom1" data="found" num="1" reduce="mindist"/>
-    <contact name="RR_calf_contact1" geom1="floor" geom2="RR_calf_geom1" data="found" num="1" reduce="mindist"/>
-
-    <contact name="FL_calf_contact2" geom1="floor" geom2="FL_calf_geom2" data="found" num="1" reduce="mindist"/>
-    <contact name="FR_calf_contact2" geom1="floor" geom2="FR_calf_geom2" data="found" num="1" reduce="mindist"/>
-    <contact name="RL_calf_contact2" geom1="floor" geom2="RL_calf_geom2" data="found" num="1" reduce="mindist"/>
-    <contact name="RR_calf_contact2" geom1="floor" geom2="RR_calf_geom2" data="found" num="1" reduce="mindist"/>
-  </sensor>
-
-</mujoco>
