+++ src/unilab/envs/locomotion/go2/__init__.py
@@ -1,4 +1,3 @@
 from .handstand import Go2HandStandCfg, Go2HandStandTask
 from .joystick import Go2JoystickCfg, Go2WalkTask
 from .rough import Go2JoystickRoughCfg, Go2JoystickRoughEnv
-from .rough_full import Go2JoystickRoughFullCfg, Go2JoystickRoughFullEnv
